Kolja Jez

+++ kolja/kolja-ansi/src/main/java/com/baulsupp/kolja/ansi/progress/NullProgressBar.java
@@ -21,6 +21,6 @@ public class NullProgressBar implements ProgressBar {
   public void clear() {
   }
 
-  public void showProgress(int completed, int total) {
+  public void showProgress(long completed, long total) {
   }
 }
